+++ packages/components/src/Chip/Chip.css.d.ts
@@ -1,0 +1,14 @@
+declare const styles: {
+  readonly "chip": string;
+  readonly "prefix": string;
+  readonly "suffix": string;
+  readonly "empty": string;
+  readonly "chipBar": string;
+  readonly "selected": string;
+  readonly "base": string;
+  readonly "invalid": string;
+  readonly "subtle": string;
+  readonly "disabled": string;
+};
+export = styles;
+