+++ packages/components/src/Combobox/ComboboxProvider.tsx
@@ -3,6 +3,7 @@ import styles from "./Combobox.css";
 
 export const ComboboxContext = React.createContext(
   {} as {
+    multiselect: boolean;
     open: boolean;
     setOpen: (open: boolean) => void;
     wrapperRef: RefObject<HTMLDivElement>;
@@ -11,16 +12,20 @@ export const ComboboxContext = React.createContext(
 
 export interface ComboboxProviderProps {
   readonly children: React.ReactNode;
+  readonly multiselect?: boolean;
 }
 
 export function ComboboxContextProvider(
   props: ComboboxProviderProps,
 ): JSX.Element {
+  const multiselect = props.multiselect || false;
   const [open, setOpen] = useState<boolean>(false);
   const wrapperRef = useRef<HTMLDivElement>(null);
 
   return (
-    <ComboboxContext.Provider value={{ open, setOpen, wrapperRef }}>
+    <ComboboxContext.Provider
+      value={{ multiselect, open, setOpen, wrapperRef }}
+    >
       <div ref={wrapperRef}>
         {open && (
           <div